+++ app/provider/app.go
@@ -141,6 +141,7 @@ var (
 				ibcproviderclient.ConsumerAdditionProposalHandler,
 				ibcproviderclient.ConsumerRemovalProposalHandler,
 				ibcproviderclient.EquivocationProposalHandler,
+				ibcproviderclient.ChangeRewardDenomsProposalHandler,
 			},
 		),
 		params.AppModuleBasic{},